requirements.txt
dagit>=0.14.8
dagster-postgres>=0.14.8
dagster-celery>=0.14.8
dagster>=0.14.8
pydantic>=1.9.0
python-magic>=0.4.25
py7zr>=0.18.5
jq>=1.2.2
splunk-sdk>=1.6.19
ujson>=5.2.0
evtx>=0.7.3
mft>=0.5.5
python-registry>=1.3.1